util: Return specific context error from `ConnWithContext.{Read,Write}()`
When context is closed because of an error, this error should also be
returned from `ConnWithContext.{Read,Write}()` because it could be used
by the caller.

This commit fixes e.g. a bug in `bisquitt-sub` where on some internal
errors, only "context canceled" error is printed on the console, instead
of the real error which caused the cancellation.