gateway: Increase `dtlsConnectTimeout` to 5 minutes
I'm increasing the DTLS handshake timeout as a supposed ugly hotfix to a
bug in the DTLS handshake we are experiencing in production. We will
investigate the problem further and solve it better in subsequent
commits.

We are experiencing a connect timeout in production:

```
[ERROR][gw] Client TLS handshake error: handshake error: context
deadline exceeded
```

This probably leads to some strange condition where the listener is not
able to accept further connections. Error we get:

```
[ERROR][gw] Client TLS handshake error: handshake error: write udp
[::]:8883->10.0.0.2:31103: sendto: operation not permitted
```

This could be somehow related to `pion/dtls`'s handshakes serialization:

```
Before /v2 Pion DTLS would handshake on Server or Client creation. This
design caused the Accept implementation to be blocking. A new connection
couldn't be accept until the previous one had finished.
```

but we have not investigated this yet.